refactor(ui): hoist pure tool/diff logic to platform-neutral modules (CODE-196)
